refactor: split page template
