+++ libs/datasource/db-service/src/services/editor-block/templates/grid.json
@@ -1,0 +1,159 @@
+{
+    "type": "group",
+    "properties": {},
+    "blocks": [
+        {
+            "type": "heading2",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "bold": true,
+                            "text": "Performance"
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "text",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "text": "CRDT tolerates higher latency and can wait longer for solving conflicts, whereas the calculation of OT in the same condition may become too overwhelming for a server to sustain."
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "heading2",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "bold": true,
+                            "text": "Extensibility"
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "text",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "text": "Because CRDT supports more data types and editor elements, it is more extensible."
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "heading1",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "text": "Conclusion"
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "grid",
+            "properties": {},
+            "blocks": [
+                {
+                    "type": "gridItem",
+                    "properties": {
+                        "gridItemWidth": "50%"
+                    },
+                    "blocks": [
+                        {
+                            "type": "text",
+                            "properties": {
+                                "text": {
+                                    "value": [
+                                        {
+                                            "text": "Collaborative algorithm is still a foreign concept for many developers. There are some introductions to it, but as to how it shall be used, there is still lack of clear explanation. I hope this article helps. If you also work for a startup company and want some suggestions, CRDT, especially Yjs, should be a better choice."
+                                        }
+                                    ]
+                                }
+                            },
+                            "blocks": []
+                        }
+                    ]
+                },
+                {
+                    "type": "gridItem",
+                    "properties": {
+                        "gridItemWidth": "50%"
+                    },
+                    "blocks": [
+                        {
+                            "type": "text",
+                            "properties": {
+                                "text": {
+                                    "value": [
+                                        {
+                                            "text": "Collaborative algorithm is still a foreign concept for many developers. There are some introductions to it, but as to how it shall be used, there is still lack of clear explanation. I hope this article helps. If you also work for a startup company and want some suggestions, CRDT, especially Yjs, should be a better choice."
+                                        }
+                                    ]
+                                }
+                            },
+                            "blocks": []
+                        }
+                    ]
+                }
+            ]
+        },
+        {
+            "type": "text",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "text": ""
+                        }
+                    ]
+                }
+            },
+            "blocks": []
+        },
+        {
+            "type": "text",
+            "properties": {
+                "text": {
+                    "value": [
+                        {
+                            "text": ""
+                        },
+                        {
+                            "type": "link",
+                            "url": "https://blog.affine.pro/",
+                            "id": "link.stubssslo0rq",
+                            "children": [
+                                {
+                                    "text": "← View all posts"
+                                }
+                            ]
+                        },
+                        {
+                            "text": ""
+                        }
+                    ]
+                }
+            }
+        }
+    ]
+}
